@@ -14,8 +14,7 @@ exports.createSuggestion = async (req, res) => {
 // 获取用户的建议（填充用户信息）
 exports.getSuggestionsByUserId = async (req, res) => {
   try {
-    const suggestions = await Suggestion.find({ user_id: req.params.userId })
-      .populate('user_id', 'username email');
+    const suggestions = await Suggestion.find({ user_id: req.params.userId });
     res.json(suggestions);
   } catch (err) {
     res.status(400).json({ error: err.message });
