@@ -2,7 +2,7 @@ const mongoose = require('mongoose');
 
 const suggestionSchema = new mongoose.Schema({
   user_id: { type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true },
-  date: { type: Date, default: Date.now },
+  date: { type: Date, default: new Date() },
   content: { type: String, required: true } 
 });
 
