@@ -1,52 +1,149 @@
-// frontend/src/components/Dashboard.js
 import React from 'react';
-import PropTypes from 'prop-types';
-import { AppBar, Tabs, Tab, Box, Typography } from '@mui/material';
-import UserProfile from './UserProfile';
+import {
+  Box,
+  Typography,
+  Grid,
+  Paper,
+  IconButton,
+  Badge,
+  // Avatar,
+  // Menu,
+  // MenuItem,
+} from '@mui/material';
+import NotificationsIcon from '@mui/icons-material/Notifications';
+// import MenuIcon from '@mui/icons-material/Menu';
+import { useNavigate } from 'react-router-dom';
+import Layout from '../components/Layout';
 
-function TabPanel(props) {
-  const { children, value, index, ...other } = props;
-  return (
-    <div role="tabpanel" hidden={value !== index} {...other}>
-      {value === index && (
-        <Box p={3}>
-          {children}
-        </Box>
-      )}
-    </div>
-  );
-}
-
-TabPanel.propTypes = {
-  children: PropTypes.node,
-  value: PropTypes.number.isRequired,
-  index: PropTypes.number.isRequired,
-};
+// Tab/Subheader Section Component
+const TabSection = () => (
+  <Box
+    sx={{
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+      px: 0.75,
+      mb: 4,
+    }}
+  >
+    <Typography variant="h5" fontWeight="bold" fontSize={28}>
+      Dashboard{' '}
+      <Typography component="span" fontWeight="normal" color="text.secondary">
+        Welcome back!
+      </Typography>
+    </Typography>
+    <Badge
+      badgeContent={1}
+      color="error"
+      sx={{
+        '& .MuiBadge-badge': {
+          top: -5,
+          right: -5,
+          transform: 'none',
+        },
+      }}
+    >
+      <IconButton
+        sx={{
+          backgroundColor: 'grey.300',
+          borderRadius: '50%',
+          p: 1,
+        }}
+      >
+        <NotificationsIcon />
+      </IconButton>
+    </Badge>
+  </Box>
+);
 
 const Dashboard = () => {
-  const [value, setValue] = React.useState(0);
-  const handleChange = (event, newValue) => setValue(newValue);
+  const navigate = useNavigate();
+
+  // Dimensions for the cards – adjust these values as needed
+  const cardDimensions = {
+    checkin: { minHeight: 150 },
+    history: { minHeight: 150 },
+    mood: { minHeight: 300 },
+    ai: { minHeight: 300 },
+  };
 
   return (
-    <div>
-      <AppBar position="static" sx={{ backgroundColor: '#e6f4df', color: 'black' }}>
-        <Tabs value={value} onChange={handleChange} aria-label="dashboard tabs">
-          <Tab label="Home" />
-          <Tab label="Check-Ins" />
-          <Tab label="Profile" />
-        </Tabs>
-      </AppBar>
+    <Layout>
+      {/* <HeaderSection /> */}
+      <TabSection />
 
-      <TabPanel value={value} index={0}>
-        <Typography>Welcome to MindCare!</Typography>
-      </TabPanel>
-      <TabPanel value={value} index={1}>
-        <Typography>Check-in tab (coming soon)</Typography>
-      </TabPanel>
-      <TabPanel value={value} index={2}>
-        <UserProfile />
-      </TabPanel>
-    </div>
+      {/* First row – Check-In & History */}
+      <Grid container spacing={3} mb={4} justifyContent="flex-start">
+        <Grid item sx={{ flex: '0 0 400px', maxWidth: '400px' }}>
+          <Paper
+            elevation={4}
+            sx={{
+              p: 2,
+              borderRadius: 4,
+              ...cardDimensions.checkin,
+              cursor: 'pointer',
+            }}
+            onClick={() => navigate('/checkin')}
+          >
+            <Typography variant="h6">
+              Check-In{' '}
+              <Typography component="span" color="error" fontWeight="bold">
+                1
+              </Typography>
+            </Typography>
+          </Paper>
+        </Grid>
+        <Grid item sx={{ flex: '0 0 400px', maxWidth: '400px' }}>
+          <Paper
+            elevation={4}
+            sx={{
+              p: 2,
+              borderRadius: 4,
+              ...cardDimensions.history,
+              cursor: 'pointer',
+            }}
+            onClick={() => navigate('/history')}
+          >
+            <Typography variant="h6">History</Typography>
+          </Paper>
+        </Grid>
+      </Grid>
+
+      <hr />
+
+      {/* Second row – Mood Trends & AI-Insights */}
+      <Grid container spacing={3} mt={2}>
+        <Grid item xs={12} md={6}>
+          <Paper
+            elevation={4}
+            sx={{
+              p: 2,
+              borderRadius: 4,
+              ...cardDimensions.mood,
+              cursor: 'pointer',
+            }}
+            onClick={() => navigate('/mood')}
+          >
+            <Typography variant="h6">Mood Trends</Typography>
+          </Paper>
+        </Grid>
+        <Grid item xs={12} md={6}>
+          <Paper
+            elevation={4}
+            sx={{
+              p: 2,
+              borderRadius: 4,
+              ...cardDimensions.ai,
+              cursor: 'pointer',
+            }}
+            onClick={() => navigate('/ai')}
+          >
+            <Typography variant="h6">AI-Insights</Typography>
+          </Paper>
+        </Grid>
+      </Grid>
+    {/* </Box> */}
+    </Layout>
   );
 };
 
